fix: fixin migrations and fatories

+++ database/migrations/2025_03_25_204448_create_expenses_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->timestamps();
             $table->text('description')->nullable();
-            $table->integer('value');
+            $table->unsignedSmallInteger('value');
             $table->softDeletes();
 
             $table->unsignedBigInteger('expense_type_id');
